Undo the server setting stuff
